generate_message::helpers

Function update_modify_encryption_specs

pub fn update_modify_encryption_specs(
    address: &str,
    specs: &mut NetworkSpecs,
    encryption: &Encryption,
    optional_signer_title_override: Option<String>,
    optional_token_override: Option<Token>,
) -> Result<()>

Modify [NetworkSpecs] existing in the database only with different encryption.

New data always will be added to the database unless errors occur.

Function inputs:

&str address to make RPC calls

NetworkSpecs as they were found in the database, to be modified here

new Encryption to apply to encryption and title (if no title override was entered) fields of the NetworkSpecs

optional title override

optional token override