generate_message

Module interpret_specs

Fit network properties fetched via RPC call into data in expected format

Fetched network properties contain data needed to make NetworkSpecs, used in ContentAddSpecs payload.

Network properties may contain:

base58 prefix for the network (to be fit into u16)

decimals to display balance-related values with correct order of magnitude (to be fit into u8)

unit to display balance-related values with correct units (to be fit into String)

There could be base58 prefix information in network metadata. If base58 prefix is fetched via system_properties RPC call and the metadata contains SS58Prefix constant, they must be matching for the network to be suitable for use in the Vault. If no base58 prefix is encountered at all, network is not suitable for use in Vault either.

Some networks support more than one token, in this case RPC returns an array of decimals and an array of units, that must have equal length. No mechanism to automatically choose a token from the fetched set is currently present.

Some other networks have no token at all.

For networks with token set or no token, decimals default to 0 and units default to UNIT. For networks with token set there could be a token override applied by user in command line sequence, to set the decimals and units manually. Token override is not supported for networks that have a single token or no token.

Command line with token override:

$ cargo run add-specs -d -u <url_address> --encryption sr25519 --token-decimals <decimals> --token-unit <unit>

Enums

TokenFetch

Accepted token from system_properties RPC call

Functions

check_specs

Get from system_properties RPC call results the network data to be compared with already known data.

interpret_properties

Transform system_properties RPC call results into [NetworkProperties].