generate_message::parser

Enum Command

pub enum Command {

    Show {
        s: Show,
        db_path: PathBuf,
    },
    Specs {
        s: InstructionSpecs,
    },
    Load(InstructionMeta),
    Types {
        db_path: PathBuf,
        files_dir: PathBuf,
    },
    Make(Make),
    Sign(Make),
    Remove {
        r: Remove,
        db_path: PathBuf,
    },
    RestoreDefaults {
        db_path: PathBuf,
    },
    MakeColdRelease {
        path: Option<PathBuf>,
    },
    TransferMetaToColdRelease {
        cold_db: PathBuf,
        hot_db: PathBuf,
    },
    Unwasm {
        filename: String,
        update_db: bool,
        db_path: PathBuf,
        files_dir: PathBuf,
    },
    MetaDefaultFile {
        name: String,
        version: u32,
        db_path: PathBuf,
        export_dir: PathBuf,
    },
    MetaAtBlock {
        url: String,
        block_hash: String,
        export_dir: PathBuf,
    },
    EncodeToQr {
        path: Option<PathBuf>,
        hex: Option<String>,
        chunk_size: u16,
        dst_file: PathBuf,
    },
    KeyInfoExportToQr {
        dst_file: PathBuf,
        chunk_size: u16,
        fps: u16,
        keys_num: usize,
    },
    BulkTransactionTestPayload {
        dst_file: PathBuf,
        tx_count: usize,
        chunk_size: u16,
        from: String,
        output_format: Goal,
    },

}

Variants

Show

Display content of the a given tree of the hot database

Fields

s: Show

db_path: PathBuf

Path to the hot database

Specs

Prepare payload for add-specs update

This will remove:

address book entry AddressBookEntry from ADDRESS_BOOK tree

network specs NetworkSpecs from SPECSTREEPREP tree

all associated metadata entries from METATREE if there are no other address book entries this metadata is associated with

all associated meta block history entries from META_HISTORY if there are no other address book entries this block history entries are associated with

Fields

s: InstructionSpecs

Load(InstructionMeta)

Prepare payload for load-metadata update

Types

Prepare payload for load-types update

Fields

db_path: PathBuf

Path to hot db

files_dir: PathBuf

Folder to save payloads ready for signing

Make(Make)

Complete update generation according

Sign(Make)

Sign

Remove

Remove all data associated with a network

This will remove:

address book entry AddressBookEntry from ADDRESS_BOOK tree

network specs NetworkSpecs from SPECSTREEPREP tree

all associated metadata entries from METATREE if there are no other address book entries this metadata is associated with

all associated meta block history entries from META_HISTORY if there are no other address book entries this block history entries are associated with

Fields

r: Remove

db_path: PathBuf

Path to the hot database

RestoreDefaults

Restore hot database to default state

Removes old hot database and generates new one with default values at default path HOT_DB_NAME.

By default, hot database contains:

ADDRESS_BOOK entries for default networks

SPECSTREEPREP entries for default networks

types information in SETTREE

no metadata entries in METATREE

no meta block history entries in META_HISTORY

Default networks are Polkadot, Kusama, and Westend.

Fields

db_path: PathBuf

Path to hot db

MakeColdRelease

Generate release cold database at optionally provided path

Removes old cold release database and generates new one with default values (uninitiated) at user-provided path or, if no valid path is given, at default path COLD_DB_NAME_RELEASE.

By default, the uninitiated cold release database contains:

SPECSTREE entries for default networks

VERIFIERS entries for default networks, with verifiers set to the general one

two latest metadata versions for default networks in METATREE

default types information and clean danger status in SETTREE

Note that the general verifier is not specified and history is not started. This will be done only in Vault itself. Before initialization, the cold release database could not be used by Vault.

Fields

path: Option<PathBuf>

Path to release db

TransferMetaToColdRelease

Transfer metadata from hot database to release cold database

Metadata from hot database is transferred to cold release database at user-provided path or, if no valid path is given, at default path COLD_DB_NAME_RELEASE.

Metadata is transferred only for the networks that are known to the cold database, i.e. the ones having OrderedNetworkSpecs entry in SPECSTREE.

Fields

cold_db: PathBuf

Path to release db

hot_db: PathBuf

Path to hot db

Unwasm

Prepare payload for load_metadata update from .wasm file

This command extracts metadata from .wasm file and uses this metadata to produce load_metadata update payload. Only networks with network specs entries in the hot database could be processed with unwasm command, since the load_metadata update payload in addition to metadata requires also network genesis hash. unwasm command could be used to generate update QR codes before the metadata becomes accessible from the node.

Network name found in the metadata is used to find NetworkSpecs for the network. NetworkSpecs are used to get genesis hash and to check base58 prefix, it the network metadata has base58 prefix inside.

A raw bytes update payload file is generated in dedicated FOLDER to (optionally) be signed and later be transformed into load_metadata update QR. Update payload file name is sign_me_load_metadata_<network_name>V<version>.

By default, metadata extracted from .wasm file is added to the database. Optional -d key could be used is database should not be updated. If the metadata gets entered in the database (i.e. no -d key used), META_HISTORY gets no entry. Block hash will be added if the same metadata is later fetched from a node.

Fields

filename: String

WASM file

update_db: bool

update the DB.

db_path: PathBuf

Hot database path

files_dir: PathBuf

Folder to save payloads ready for signing

MetaDefaultFile

Make file with hexadecimal metadata for defaults release metadata set

Produces file with hex-encoded network metadata from the hot database METATREE entry.

Output file named <network_name><metadata_version> is generated in dedicated EXPORT_FOLDER. It contains hexadecimal network metadata.

Fields

name: String

File name

version: u32

Version

db_path: PathBuf

Hot database path

export_dir: PathBuf

Folder to save completed update messages

MetaAtBlock

Create file with network metadata at block hash

Output file named <network_name><metadata_version>_<block_hash> is generated in dedicated folder. It contains hexadecimal network metadata. This command does not address or update the hot database.

Fields

url: String

URL of the chain RPC point

block_hash: String

Hash of the block at which meta is asked

export_dir: PathBuf

Folder to save completed update messages

EncodeToQr

Encode payload to multiframe QR

Fields

path: Option<PathBuf>

Path to a file to encode

hex: Option<String>

Hex-encoded payload to encode

chunk_size: u16

Size of a chunk in a multiframe RaptorQ encoding

dst_file: PathBuf

Destination file to write qr code to

KeyInfoExportToQr

Produce a test key info export QR

Fields

dst_file: PathBuf

chunk_size: u16

Size of a chunk in a multiframe RaptorQ encoding.

fps: u16

FPS of a multiframe RaptorQ encoding

keys_num: usize

Number of keys to generate and export

BulkTransactionTestPayload

Produce a bulk tx signing QR

Fields

dst_file: PathBuf

tx_count: usize

Number of transactions in a bulk.

chunk_size: u16

Size of a chunk in a multiframe RaptorQ encoding.

from: String

AccountId of transaction in hex form.

output_format: Goal

Qr or File