Enum generate_message::Error

pub enum Error {

    Database(Error),
    Io(Error),
    Definitions(Error),
    Sled(Error),
    JsonRPC(Error),
    Codec(Error),
    Specs(SpecsError),
    Metadata(MetadataError),
    Qr(Box<dyn Error>),
    SignatureWrongLength(usize, usize),
    PublicKeyWrongLength(usize, usize),
    SpecsInDb {
        name: String,
        encryption: Encryption,
    },
    NotFound(String),
    NotSupported,
    TwoGenesisHashVariantsForName {
        name: String,
    },
    TwoUrlVariantsForName {
        name: String,
    },
    TwoBase58ForName {
        name: String,
    },
    AddressBookEmpty,
    UKeyUrlInDb {
        title: String,
        url: String,
    },
    UKeyHashInDb {
        address_book_entry: AddressBookEntry,
        url: String,
    },
    UnexpectedMetadataFormat,
    UnexpectedGenesisHashFormat,
    UnexpectedSystemPropertiesFormat,
    UnexpectedBlockHashFormat,
    AddressBookSpecsName {
        address_book_name: String,
        specs_name: String,
    },
    NetworkSpecs(NetworkSpecsKey),
    TwoNamesForUrl {
        url: String,
    },
    HotDatabaseMetadataOverTwoEntries {
        name: String,
    },
    HotDatabaseMetadataSameVersionTwice {
        name: String,
        version: u32,
    },
    EarlierVersion {
        name: String,
        old_version: u32,
        new_version: u32,
    },
    SameVersionDifferentMetadata {
        name: String,
        version: u32,
        block_hash_in_db: Option<H256>,
        block_hash_in_fetch: Option<H256>,
    },
    ValuesChanged {
        url: String,
        what: Changed,
    },
    FaultySpecs {
        url: String,
        error: SpecsError,
    },
    UnexpectedFetchedGenesisHashFormat {
        value: String,
    },
    UnexpectedFetchedBlockHashFormat {
        value: String,
    },
    BlockHashLength,
    AddressBookEntryWithName {
        name: String,
    },
    BadSignature,
    UnexpectedPubKeyLength,
    UnexpectedSignatureLength,

}

Generate Message error.

Variants

Database(Error)

Io(Error)

Definitions(Error)

Sled(Error)

JsonRPC(Error)

Codec(Error)

Specs(SpecsError)

Metadata(MetadataError)

Qr(Box<dyn Error>)

SignatureWrongLength(usize, usize)

PublicKeyWrongLength(usize, usize)

SpecsInDb

Fields

name: String

network name

encryption: Encryption

network supported encryption

Network specs are already in the database

NotFound(String)

NotSupported

TwoGenesisHashVariantsForName

Fields

name: String

network name

ADDRESS_BOOK tree of the hot database contains AddressBookEntry entries with same name field and different genesis_hash values.

This is not allowed, as it would cause uncertainty in load_metadata message generation, which is build using metadata and genesis hash.

TwoUrlVariantsForName

Fields

name: String

network name

ADDRESS_BOOK tree of the hot database contains AddressBookEntry entries with same name field and different address values.

Hot database does not allow to store more than one trusted URL address for RPC calls for same network.

Alternative URL address could be used if the database is not updated (-d key is used).

To update the address in the database in case the old one is no longer acceptable, one should remove old entry, and only then add the new one.

TwoBase58ForName

Fields

name: String

network name

SPECSTREEPREP tree of the hot database contains NetworkSpecs entries with same network name, but different base58 prefix.

AddressBookEmpty

ADDRESS_BOOK tree of the hot database has no entries to process.

UKeyUrlInDb

Fields

title: String

network address book title

url: String

URL address

Tried to fetch with -u key using address already known to the database

UKeyHashInDb

Fields

address_book_entry: AddressBookEntry

address book entry with exactly matching genesis hash

url: String

URL address used for fetch

Tried to fetch with -u key using address not known to the database, but got genesis hash that is already known.

Likely tried to fetch with different address when one already is in the database.

Hot database does not allow to store more than one trusted URL address for RPC calls for same network.

Alternative URL address could be used if the database is not updated (-d key is used).

To update the address in the database in case the old one is no longer acceptable, one should remove old entry, and only then add the new one.

UnexpectedMetadataFormat

UnexpectedGenesisHashFormat

UnexpectedSystemPropertiesFormat

UnexpectedBlockHashFormat

AddressBookSpecsName

Fields

address_book_name: String

name in AddressBookEntry

specs_name: String

name in NetworkSpecs

AddressBookEntry in hot database contains encryption and genesis_hash fields, from which the corresponding [NetworkSpecsKey] could be built.

NetworkSpecsKey has an associated NetworkSpecs value stored in SPECSTREEPREP tree of the hot database. NetworkSpecs has field name with network name.

This error appears if the name from NetworkSpecs differs from the name in AddressBookEntry.

NetworkSpecs(NetworkSpecsKey)

TwoNamesForUrl

Fields

url: String

URL address, for which two condlicting names were found

ADDRESS_BOOK tree of the hot database contains AddressBookEntry entries with same address field and different name fields.

Name in address book is taken from the metadata, metadata is fetched using RPC call, so one URL address can correspond to only one network name.

HotDatabaseMetadataOverTwoEntries

Fields

name: String

network name

METATREE of the hot database should contain at most two latest metadata versions for each network, with the older entries being deleted as the new ones appear.

This error appears if during the processing more than two metadata entries for the network are found.

HotDatabaseMetadataSameVersionTwice

Fields

name: String

network name

version: u32

network version

METATREE of the hot database has two entries for a network with the same metadata version.

Note: at this moment should be unreachable, since the entries are getting checked for consistency with MetaKey.

EarlierVersion

Fields

name: String

network name

old_version: u32

network version in hot database, higher than the one just fetched

new_version: u32

network version just fetched

Fetched network metadata version is lower than the one already in the hot database.

SameVersionDifferentMetadata

Fields

name: String

network name

version: u32

network version

block_hash_in_db: Option<H256>

optionally recorded block hash for which the metadata was fetched when recorded in the database

block_hash_in_fetch: Option<H256>

block hash for which the metadata is fetched now

Fetched network metadata is different from the one already in the hot database, with the same network name and network version.

ValuesChanged

Fields

url: String

URL address used for RPC call

what: Changed

what exactly has changed

Fetched data is different from the one already in the hot database.

FaultySpecs

Fields

url: String

URL address used for RPC cal

error: SpecsError

what exactly is wrong with the network specs

Fetched network specs are not suitable for use in Vault.

UnexpectedFetchedGenesisHashFormat

Fields

value: String

genesis hash value as received through RPC call

Fetched genesis hash could not be transformed in expected [u8; 32] value.

UnexpectedFetchedBlockHashFormat

Fields

value: String

block hash value as received through RPC call

Fetched block hash could not be transformed in expected [u8; 32] value.

BlockHashLength

User-entered block hash has invalid length

AddressBookEntryWithName

Fields

name: String

AddressBookEntry searched in ADDRESS_BOOK tree of the hot database by matching the name field.

Associated data is the network name used for searching.

BadSignature

Provided data signature (entered separately or as a part of SufficientCrypto input) is invalid for given public key and data.

UnexpectedPubKeyLength

UnexpectedSignatureLength