Module generate_message::fetch_metadata

Fetch network information from a node using rpc calls

Preparing add_specs and load_metadata update payload may require gathering network information from a node.

For add_specs update payload:

call 	fetched information

state_getMetadata, for current block 	current block network metadata, that will be used to get:
- network name
- base58 prefix from metadata

chain_getBlockHash, for 0th block 	network genesis hash

system_properties 	- base58 prefix
- decimals
- unit

Network name expected to remain the same for the network over time. The only way to get network name is from the network metadata Version constant.

For load_metadata update:

call 	fetched information

chain_getBlockHash, for current block 	current block hash

state_getMetadata, for just fetched block hash 	latest network metadata

chain_getBlockHash, for 0th block 	network genesis hash

Block hash is fetched first to always have network metadata matching the block hash, even if the two rpc calls were done during block switching.

Addresses for rpc calls in different networks could be found here

This module deals only with the rpc calls part and does no processing of the fetched data.

Structs

FetchedInfo

Data from rpc calls for load_metadata update payload.

FetchedInfoWithNetworkSpecs

Data from rpc calls for add_specs update payload.

Functions

fetch_info

Fetch data for load_metadata update payload through rpc calls.

fetch_info_with_network_specs

Fetch data for add_specs update payload through rpc calls.

fetch_meta_at_block

Fetch network metadata from given url address at given block through rpc call.