Module generate_message::helpers

Helpers

Structs

MetaFetched

Fetched and interpreted data for load_metadata payload and database update.

MetaShortCut

Data needed to output load_metadata update payload file.

MetaValuesStamped

MetaValues with corresponding block hash at the time of fetch, if available.

SortedMetaValues

MetaValuesStamped sorted into sets of newer and older, by metadata version.

Enums

Write

Content to print during load_metadata <-k/-p/-t> processing.

Functions

add_new_metadata

Try updating SortedMetaValues with new MetaValuesStamped.

add_specs_print

Write to file add_specs update payload as raw bytes.

address_book_content

Get all ADDRESS_BOOK entries with address book titles.

db_upd_metadata

Update the database after load_metadata run.

db_upd_network

Update the database after add_specs run.

debug_meta_at_block

Get network metadata file from given url address at specified block.

error_occured

Process error depending on pass errors flag -s.

filter_address_book_by_url

Get all ADDRESS_BOOK entries with address book titles, for given url address.

genesis_hash_in_hot_db

Search for any ADDRESS_BOOK entry with given genesis hash.

get_address_book_entry

Get AddressBookEntry from the database for given address book title.

get_network_specs_to_send

Get network specs NetworkSpecsToSend from the hot database.

is_specname_in_db

Check if ADDRESS_BOOK has entries with given name and title other than except_title.

load_metadata_print

Write to file load_metadata update payload as raw bytes.

meta_fetch

Get network information through rpc calls at address and interpret it into MetaFetched.

meta_history_content

Get all entries from META_HISTORY.

network_specs_from_entry

Get NetworkSpecsToSend corresponding to the given AddressBookEntry.

network_specs_from_title

Get NetworkSpecsToSend from the database for given address book title.

prepare_metadata

Collect and sort MetaValuesStamped from the hot database

read_metadata_database

Collect all MetaValuesStamped from the hot database.

specs_agnostic

Fetch data and assemble NetworkSpecsToSend with only url address and user-entered data.

try_get_network_specs_to_send

Try to get network specs NetworkSpecsToSend from the hot database.

update_known_specs

Update NetworkSpecsToSend already existing in the database with exactly same encryption.

update_modify_encryption_specs

Modify NetworkSpecsToSend existing in the database only with different encryption.