generate_message::helpers

Function add_new_metadata

pub fn add_new_metadata(
    new: &MetaValuesStamped,
    sorted: &mut SortedMetaValues,
) -> Result<bool>

Try updating SortedMetaValues with new MetaValuesStamped.

Outputs flag to indicate that the SortedMetaValues got updated.

If the fetched metadata is good and has later version than the ones in SortedMetaValues, it is added to newer set, any previous value from newer is moved to older. If there was any value in older, it gets kicked out.

If there was no block hash in hot database and the metadata did not change, a new block hash could be added if it is known.