Function generate_message::interpret_specs::check_specs

pub fn check_specs(
    x: &Map<String, Value>, 
    optional_prefix_from_meta: Option<u16>
) -> Result<(u16, TokenFetch), SpecsError>

Get from system_properties rpc call results the network data to be compared with already known data.

This function is used if the fetch results are used to check already existing database entry.

Function inputs:

&Map<String, Value> received via system_properties rpc call,

optional base58 prefix from the network metadata

Function outputs:

network base58 prefix

TokenFetch